dom/midi/tests/test_midi_permission_gated.html

<html>
  <head>
    <title>WebMIDI Listener Test</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="MIDITestUtils.js"></script>
  </head>

  <body onload="runTests()">
    <iframe id="ifr"></iframe>
    <script class="testbody" type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    // Generally this runs on example.com but with --enable-xorigin-tests it runs
    // on example.org.
    $("ifr").src = "https://test1." + location.host + "/tests/dom/midi/tests/file_empty.html";

    async function runTests() {
      await SpecialPowers.pushPrefEnv({
        set: [["dom.webmidi.enabled", true]],
      });
      ok(
        await SpecialPowers.testPermission(
          "midi-sysex",
          SpecialPowers.Services.perms.UNKNOWN_ACTION,
          document
        ),
        "midi-sysex value should have UNKNOWN permission"
      );
      ok(
        await SpecialPowers.testPermission(
          "midi-sysex",
          SpecialPowers.Services.perms.UNKNOWN_ACTION,
          SpecialPowers.wrap($("ifr")).contentDocument
        ),
        "permission should also not be set for iframe"
      );

      // Gated permission should fail without addon installed.
      for (let sysex of [false, true]) {
        try {
          await navigator.requestMIDIAccess({ sysex });
          ok(false, "MIDI Access Request gate failed");
        } catch (ex) {
          ok(true, "MIDI Access Request denied by default");
        }
      }

      // When an addon is installed, the permission is inserted.  Test
      // that the request succeeds after we insert the permission.
      await SpecialPowers.addPermission(
        "midi-sysex",
        SpecialPowers.Services.perms.ALLOW_ACTION,
        document
      );

      // Gated permission should allow access after addon inserted permission.
      for (let sysex of [false, true]) {
        try {
          await navigator.requestMIDIAccess({ sysex });
          ok(true, "MIDI Access Request allowed");
        } catch (ex) {
          ok(false, "MIDI Access Request failed");
        }
      }

      // Gated permission should also apply to subdomains.
      for (let sysex of [false, true]) {
        try {
          await SpecialPowers.wrap($("ifr")).contentWindow.navigator.requestMIDIAccess({ sysex });
          ok(true, "MIDI Access Request allowed for subdomain");
        } catch (ex) {
          ok(false, "MIDI Access Request failed for subdomain");
        }
      }

      // Remove the permission.
      await SpecialPowers.removePermission("midi-sysex", document);

      SimpleTest.finish();
    }
    </script>
  </body>
</html>
